Added Bombs ability, which locks on to enemy units and then deals damage on activation to the target unit. It raycasts in a rough cone underneath/ahead of the bomber, looking ahead based on the velocity of the bomber.  Because it is a work-in-progress, on activation it immediately deals a large amount of damage. Added a chargeup sound. Tweaked Entity and Silhouette shaders to respect aspect ratio in their comic shading and use halfs instead of floats, respectively.

+++ Assets/Shaders/Silhouette.shader
@@ -39,7 +39,7 @@
 		float a = 1;
 		float b = 1;
 		float aspectRatio = _ScreenParams.x / _ScreenParams.y;
-		float dxy = dot(float2(co.x, co.y / aspectRatio), half2(a, b));
+		float dxy = dot(half2(co.x, co.y / aspectRatio), half2(a, b));
 		return frac(dxy * rate);
 	}
